# magazine_api/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.Transaction:
    properties:
      bank_payment_from:
        type: string
      bank_payment_to:
        type: string
      bank_payment_transaction_id:
        type: string
      created_by:
        type: string
      created_on:
        type: string
      creator_name:
        type: string
      credit_amount:
        type: number
      debit_amount:
        type: number
      deleted_by:
        type: string
      deleted_on:
        type: string
      deletor_name:
        type: string
      id:
        type: string
      online_payment_from:
        type: string
      online_payment_name:
        type: string
      online_payment_to:
        type: string
      online_payment_transaction_id:
        type: string
      paid_medium:
        type: integer
      payment_date:
        type: string
      payment_from:
        type: string
      payment_month:
        type: integer
      payment_to:
        type: string
      payment_type:
        type: string
      remarks:
        type: string
      title:
        type: string
      transaction_cost:
        type: number
      updated_by:
        type: string
      updated_on:
        type: string
      updator_name:
        type: string
    type: object
  models.TransactionBase:
    properties:
      bank_payment_from:
        type: string
      bank_payment_to:
        type: string
      bank_payment_transaction_id:
        type: string
      credit_amount:
        type: number
      debit_amount:
        type: number
      online_payment_from:
        type: string
      online_payment_name:
        type: string
      online_payment_to:
        type: string
      online_payment_transaction_id:
        type: string
      paid_medium:
        type: integer
      payment_date:
        type: string
      payment_from:
        type: string
      payment_month:
        type: integer
      payment_to:
        type: string
      payment_type:
        type: string
      remarks:
        type: string
      title:
        type: string
      transaction_cost:
        type: number
    type: object
...
host: localhost:5005
info:
  contact:
    name: 21557870 - Suman Chhetri
  description: This is an API documentation of Magazine API
  title: Magazine API Documentation
  version: v1
paths:
...
  /transaction:
    get:
      description: List assigns for cutting.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - type: object
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Transaction'
                  type: array
              type: object
      summary: Lists Transaction
      tags:
      - Transaction
    post:
      consumes:
      - application/json
      description: It creates assign for cutting.
      parameters:
      - description: Add Transaction
        in: body
        name: transaction
        required: true
        schema:
          $ref: '#/definitions/models.TransactionBase'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - type: object
            - properties:
                data:
                  $ref: '#/definitions/models.Transaction'
              type: object
      summary: Create Transaction
      tags:
      - Transaction
  /transaction/{id}:
    delete:
      description: Delete by Transaction ID
      parameters:
      - description: assign Cutting ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            allOf:
            - type: object
            - properties:
                data:
                  type: string
              type: object
      summary: Soft Delete Transaction
      tags:
      - Transaction
    patch:
      consumes:
      - application/json
      description: Updates Transaction
      parameters:
      - description: Update Transaction
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - type: object
            - properties:
                data:
                  $ref: '#/definitions/models.Transaction'
              type: object
      summary: Update Transaction
      tags:
      - Transaction
  /transaction/date/{date}:
    get:
      description: Gets All Transactions by exact date
      parameters:
      - description: Date
        in: path
        name: date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - type: object
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Transaction'
                  type: array
              type: object
      summary: Gets Transactions by date
      tags:
      - Transaction
...
  /transaction/forcedelete/{id}:
    delete:
      description: Delete by Transaction ID
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            allOf:
            - type: object
            - properties:
                data:
                  type: string
              type: object
      summary: Permenant Delete an Transaction
      tags:
      - Transaction
  /transaction/fromaccount/{account}:
    get:
      description: Gets All Transactions done from Account.
      parameters:
      - description: ID
        in: path
        name: account
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - type: object
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Transaction'
                  type: array
              type: object
      summary: Gets Transactions by Account.
      tags:
      - Transaction
  /transaction/id/{id}:
    get:
      description: Gets One Transactions by ID
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - type: object
            - properties:
                data:
                  $ref: '#/definitions/models.Transaction'
              type: object
      summary: Gets One Transaction by ID
      tags:
      - Transaction
  /transaction/medium/{paid_medium}:
    get:
      description: Gets All Transactions by Paid Medium
      parameters:
      - description: Type
        in: path
        name: paid_medium
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - type: object
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Transaction'
                  type: array
              type: object
      summary: Gets Transactions by Medium
      tags:
      - Transaction
  /transaction/month/{month}:
    get:
      description: Gets All Transactions by month
      parameters:
      - description: Month
        in: path
        name: month
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - type: object
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Transaction'
                  type: array
              type: object
      summary: Gets Transactions by month
      tags:
      - Transaction
  /transaction/toaccount/{account}:
    get:
      description: Gets All Transactions done to the Account.
      parameters:
      - description: ID
        in: path
        name: account
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - type: object
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Transaction'
                  type: array
              type: object
      summary: Gets Transactions to Account.
      tags:
      - Transaction
  /transaction/type/{paid_type}:
    get:
      description: Gets all Transactions by Type
      parameters:
      - description: Type
        in: path
        name: paid_type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - type: object
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Transaction'
                  type: array
              type: object
      summary: Gets Transaction by Type
      tags:
      - Transaction
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"